Refactor to packages
