+++ src/main/java/secondWeek/Ordersystem/model/Order.java
@@ -1,4 +1,4 @@
-package secondWeek.Ordersystem;
+package secondWeek.Ordersystem.model;
 
 import java.util.List;
 import java.util.Objects;
