+++ src/main/java/secondWeek/Ordersystem/services/OrderService.java
@@ -1,4 +1,9 @@
-package secondWeek.Ordersystem;
+package secondWeek.Ordersystem.services;
+
+import secondWeek.Ordersystem.db.OrderDb;
+import secondWeek.Ordersystem.db.ProductDb;
+import secondWeek.Ordersystem.model.Order;
+import secondWeek.Ordersystem.model.Product;
 
 import java.util.List;
 import java.util.Optional;
@@ -6,12 +11,12 @@ import java.util.Optional;
 public class OrderService {
  private final OrderDb orderDb;
  private final ProductDb productDb;
- private final Storage storage;
+ private final StorageService storage;
 
  public OrderService () {
 	this.productDb = new ProductDb();
 	this.orderDb = new OrderDb();
-	this.storage = new Storage();
+	this.storage = new StorageService();
  }
 
  public void addOrder (Order order) {
