+++ src/test/java/secondWeek/Ordersystem/OrderDbTester.java
@@ -1,6 +1,9 @@
 package secondWeek.Ordersystem;
 
 import org.junit.jupiter.api.Test;
+import secondWeek.Ordersystem.db.OrderDb;
+import secondWeek.Ordersystem.model.Order;
+import secondWeek.Ordersystem.model.Product;
 
 import java.util.ArrayList;
 import java.util.List;
