+++ src/test/java/secondWeek/Ordersystem/ProductDbTester.java
@@ -1,6 +1,8 @@
 package secondWeek.Ordersystem;
 
 import org.junit.jupiter.api.Test;
+import secondWeek.Ordersystem.db.ProductDb;
+import secondWeek.Ordersystem.model.Product;
 
 import java.util.ArrayList;
 import java.util.List;
