+++ src/test/java/secondWeek/Ordersystem/StorageServiceTester.java
@@ -1,6 +1,11 @@
 package secondWeek.Ordersystem;
 
 import org.junit.jupiter.api.Test;
+import secondWeek.Ordersystem.db.ProductDb;
+import secondWeek.Ordersystem.model.Order;
+import secondWeek.Ordersystem.model.Product;
+import secondWeek.Ordersystem.services.OrderService;
+import secondWeek.Ordersystem.services.StorageService;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -9,7 +14,7 @@ import java.util.Optional;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class StorageTester {
+public class StorageServiceTester {
 
  @Test
  void putAndGetOrderDb () {
@@ -32,7 +37,7 @@ public class StorageTester {
 	orderService.addOrder( order1 );
 	orderService.addOrder( order2 );
 
-	Storage storage = new Storage();
+	StorageService storage = new StorageService();
 	Optional<List<Order>> actual = storage.refreshOrders( orderService.listOrders() ); // optional
 
 	Product expectedOrder1Product1 = new Product( 1, "Roggenvollkornbrot", true );
@@ -59,7 +64,7 @@ public class StorageTester {
 
  @Test
  void putAndGetOrderDbEmpty () {
-	Storage storage = new Storage();
+	StorageService storage = new StorageService();
 	OrderService orderService = new OrderService();
 	Optional<List<Order>> actual = storage.refreshOrders( orderService.listOrders() );
 	assertTrue( actual.isEmpty() );
@@ -77,7 +82,7 @@ public class StorageTester {
 	productDb.add( orderProduct3 );
 	productDb.add( orderProduct4 );
 
-	Storage storage = new Storage();
+	StorageService storage = new StorageService();
 	Optional<List<Product>> actual = storage.refreshProducts( productDb.list() );
 
 	Product expectedOrderProduct1 = new Product( 1, "Roggenvollkornbrot", true );
@@ -97,7 +102,7 @@ public class StorageTester {
  @Test
  void putAndGetProductDbEmpty () {
 	ProductDb productDb = new ProductDb();
-	Storage storage = new Storage();
+	StorageService storage = new StorageService();
 	Optional<List<Product>> actual = storage.refreshProducts( productDb.list() );
 	assertTrue( actual.isEmpty() );
  }
